+++ annotations/src/main/java/react4j/annotations/Prop.java
@@ -16,6 +16,7 @@ import javax.annotation.Nonnull;
  * <li>Must return a value</li>
  * <li>Must be an abstract instance method</li>
  * <li>Must not throw exceptions</li>
+ * <li>Must be accessible from the same package as the class annotated by {@link ReactComponent}</li>
  * </ul>
  */
 @Documented