+++ processor/src/main/java/react4j/processor/MethodChecks.java
@@ -1,9 +1,15 @@
 package react4j.processor;
 
+import java.util.Objects;
+import java.util.Set;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
+import javax.lang.model.element.Name;
+import javax.lang.model.element.PackageElement;
+import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeKind;
 
 final class MethodChecks
@@ -22,6 +28,16 @@ final class MethodChecks
     }
   }
 
+  static void mustNotBeAbstract( @Nonnull final String annotationName, @Nonnull final ExecutableElement method )
+    throws ReactProcessorException
+  {
+    if ( method.getModifiers().contains( Modifier.ABSTRACT ) )
+    {
+      throw new ReactProcessorException( "@" + ProcessorUtil.toSimpleName( annotationName ) +
+                                         " target must not be abstract", method );
+    }
+  }
+
   static void mustBeStatic( @Nonnull final String annotationName, @Nonnull final Element method )
     throws ReactProcessorException
   {
@@ -29,6 +45,16 @@ final class MethodChecks
     {
       throw new ReactProcessorException( "@" + ProcessorUtil.toSimpleName( annotationName ) +
                                          " target must be static", method );
+    }
+  }
+
+  private static void mustNotBeStatic( @Nonnull final String annotationName, @Nonnull final Element method )
+    throws ReactProcessorException
+  {
+    if ( method.getModifiers().contains( Modifier.STATIC ) )
+    {
+      throw new ReactProcessorException( "@" + ProcessorUtil.toSimpleName( annotationName ) +
+                                         " target must not be static", method );
     }
   }
 
@@ -90,5 +116,60 @@ final class MethodChecks
       throw new ReactProcessorException( "@" + ProcessorUtil.toSimpleName( annotationName ) +
                                          " target must not throw any exceptions", method );
     }
+  }
+
+  /**
+   * Verifies that the method is not static, abstract or private.
+   * The intent is to verify that it can be instance called by sub-class in same package.
+   */
+  static void mustBeSubclassCallable( @Nonnull final TypeElement targetType,
+                                      @Nonnull final String annotationName,
+                                      @Nonnull final ExecutableElement method )
+    throws ReactProcessorException
+  {
+    mustNotBeStatic( annotationName, method );
+    mustNotBePrivate( annotationName, method );
+    mustNotBePackageAccessInDifferentPackage( targetType, annotationName, method );
+  }
+
+  static void mustNotBePackageAccessInDifferentPackage( @Nonnull final TypeElement component,
+                                                        @Nonnull final String annotationName,
+                                                        @Nonnull final ExecutableElement method )
+    throws ReactProcessorException
+  {
+    final Set<Modifier> modifiers = method.getModifiers();
+    final boolean isPackageAccess =
+      !modifiers.contains( Modifier.PRIVATE ) &&
+      !modifiers.contains( Modifier.PROTECTED ) &&
+      !modifiers.contains( Modifier.PUBLIC );
+
+    if ( isPackageAccess )
+    {
+      final PackageElement packageElement = getPackageElement( component );
+      final PackageElement methodPackageElement = getPackageElement( (TypeElement) method.getEnclosingElement() );
+      final Name componentPackageName = null == packageElement ? null : packageElement.getQualifiedName();
+      final Name methodPackageName = null == methodPackageElement ? null : methodPackageElement.getQualifiedName();
+      if ( !Objects.equals( componentPackageName, methodPackageName ) )
+      {
+        throw new ReactProcessorException( "@" + ProcessorUtil.toSimpleName( annotationName ) +
+                                           " target must not be package access if the method is in a different " +
+                                           "package from the @ArezComponent", method );
+      }
+    }
+  }
+
+  @Nullable
+  private static PackageElement getPackageElement( @Nonnull final TypeElement element )
+  {
+    Element enclosingElement = element.getEnclosingElement();
+    while ( null != enclosingElement )
+    {
+      if ( enclosingElement instanceof PackageElement )
+      {
+        return (PackageElement) enclosingElement;
+      }
+      enclosingElement = enclosingElement.getEnclosingElement();
+    }
+    return null;
   }
 }