+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseCallbackModel.java
@@ -1,0 +1,13 @@
+package com.example.package_access.other;
+
+import react4j.annotations.Callback;
+import react4j.core.Component;
+
+public abstract class BaseCallbackModel
+  extends Component
+{
+  @Callback
+  void handleFoo()
+  {
+  }
+}