+++ processor/src/test/resources/bad_input/com/example/package_access/other/BaseStateModel.java
@@ -1,0 +1,13 @@
+package com.example.package_access.other;
+
+import react4j.annotations.State;
+import react4j.core.Component;
+
+public abstract class BaseStateModel
+  extends Component
+{
+  @State
+  abstract String getFoo();
+
+  abstract void setFoo( String v );
+}