+++ processor/src/test/resources/expected/com/example/callback/FinalCallback_.java
@@ -1,0 +1,16 @@
+package com.example.callback;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import react4j.annotations.Callback;
+
+@Generated("react4j.processor.ReactProcessor")
+class FinalCallback_ {
+  private FinalCallback_() {
+  }
+
+  @Nonnull
+  static Callback.Procedure _handleFoo(@Nonnull final FinalCallback component) {
+    return ((React4j_FinalCallback) component)._handleFoo;
+  }
+}